Rename installer namespace to ExpressiveInstaller
- per conversation with @akrabat
