+++ src/ExpressiveInstaller/Resources/config/error-handler-whoops.php
