+++ src/ExpressiveInstaller/Resources/config/routes-minimal-aura-router.php
