+++ src/ExpressiveInstaller/Resources/config/routes-minimal-zf2-router.php
