+++ src/ExpressiveInstaller/Resources/templates/plates-home-page.phtml
