+++ src/ExpressiveInstaller/Resources/templates/zend-view-home-page.phtml
